moved remote corfu table key operations to utils package for availability to both server and client
