+++ infrastructure/src/test/java/org/corfudb/infrastructure/remotecorfutable/DatabaseHandlerTest.java
@@ -4,8 +4,10 @@ import com.google.common.primitives.Bytes;
 import com.google.common.primitives.Longs;
 import com.google.protobuf.ByteString;
 import lombok.extern.slf4j.Slf4j;
-import static org.corfudb.infrastructure.remotecorfutable.utils.DatabaseConstants.DATABASE_CHARSET;
-import static org.corfudb.infrastructure.remotecorfutable.utils.DatabaseConstants.EMPTY_VALUE;
+import static org.corfudb.utils.remotecorfutable.DatabaseConstants.DATABASE_CHARSET;
+import static org.corfudb.utils.remotecorfutable.DatabaseConstants.EMPTY_VALUE;
+import org.corfudb.utils.remotecorfutable.RemoteCorfuTableEntry;
+import org.corfudb.utils.remotecorfutable.RemoteCorfuTableVersionedKey;
 import org.junit.After;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
