+++ utils/src/main/java/org/corfudb/utils/remotecorfutable/KeyEncodingUtil.java
@@ -1,4 +1,4 @@
-package org.corfudb.infrastructure.remotecorfutable.utils;
+package org.corfudb.utils.remotecorfutable;
 
 import com.google.protobuf.ByteString;
